adjust amount field's script trigger to not change the rate unless it needs to
As best I can tell, this closes #51. All the calcs match what the QBO UI does.
QBO still some something strange with the line amount response from the API
where I enter exactly 2.37, but the API response shows the amount as
2.3700000000000001. I haven't found any documentation stating why it does that.

The current jsonGet function returns 2.37 if the JSON contains the above
number, so a different json parsing method might return the number with it's
full presision, which is something to be aware of. To be on the safe side,
I'm now rounding the value in the sync scripts.